refactor: share component - AccordionSummaryWithCheckbox
